% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianDeconvolution.R
\name{deconv_gauss}
\alias{deconv_gauss}
\title{Gaussian Deconvolution}
\usage{
deconv_gauss(imageMatrix, sig = 10, kern.trunc = 0.05, nw = 4)
}
\arguments{
\item{imageMatrix}{a imported image, can be imported with tiff_import()}

\item{sig}{significance parameter, used in dnorm as the final quantile}

\item{kern.trunc}{Truncated value for the kernel}

\item{nw}{A positive double-precision number, the time-bandwidth parameter}
}
\value{
A gaussian deconvoluted image matrix
}
\description{
does a 1D deconvolution on a selected image matrix
}
