% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{.env_start_end}
\alias{.env_start_end}
\title{Envelope Starts and Ends}
\usage{
.env_start_end(traceMatrix, thresh = 300)
}
\arguments{
\item{traceMatrix}{a single isolated trace, with only one trace(no timing on the plot)}

\item{thresh}{How far it will go before deciding were already past the numbers at start of traces}
}
\value{
list of the Start and End
}
\description{
A better version of .get_image_starts_ends but meant for single traces that
are isolated on their own picture.
}
