% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{.find_bounds}
\alias{.find_bounds}
\title{Funding Rough Bounds for Two Traces}
\usage{
.find_bounds(imageMatrix, FindPeaksdf, rowSums)
}
\arguments{
\item{imageMatrix}{An imported image, can be imported with tiff_import()}

\item{FindPeaksdf}{-- Row Index, Max Height, PeakStart, PeakEnd -- identifies traces by summing rows}

\item{rowSums}{Of the imported matrix from original picture}
}
\value{

}
\description{
Funding Rough Bounds for Two Traces
}
