% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{.find_bounds_2}
\alias{.find_bounds_2}
\title{Finding Rough Bounds for Two Traces}
\usage{
.find_bounds_2(imageMatrix, FindPeaksdf, rowSums, max_roc = 20)
}
\arguments{
\item{imageMatrix}{An imported image, can be imported with tiff_import()}

\item{FindPeaksdf}{-- Row Index, Max Height, PeakStart, PeakEnd -- identifies traces by summing rows}

\item{rowSums}{Of the imported matrix from original picture}

\item{max_roc}{Maximum Rate-of-Change of pixel location column-by-column}
}
\value{

}
\description{
Finding Rough Bounds for Two Traces
}
