% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{.for_bright_image}
\alias{.for_bright_image}
\title{Bright image scaling to gaussian}
\usage{
.for_bright_image(
  imageMatrix,
  sig = 10,
  kern.trunc = 0.05,
  nw = 3,
  brightQuantile = 0.95
)
}
\arguments{
\item{imageMatrix}{An imported image, can be imported with tiff_import()}

\item{sig}{Significance parameter, used in dnorm as the final quantile}

\item{kern.trunc}{Truncated value for the kernel}

\item{nw}{A positive double-precision number, the time-bandwidth parameter}

\item{brightQuantile}{the quantile that will be a cutoff for image pixels set to 0}
}
\value{
Gaussian matrix scaled for bright
}
\description{
Bright image scaling to gaussian
}
