% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{.not_bright_image}
\alias{.not_bright_image}
\title{Non bright image scaling with gaussian deconvolution}
\usage{
.not_bright_image(
  imageMatrix,
  cutoffQuantile,
  sig = 10,
  kern.trunc = 0.05,
  nw = 3
)
}
\arguments{
\item{imageMatrix}{An imported image, can be imported with tiff_import()}

\item{cutoffQuantile}{the quantile that will be a cutoff for image pixels set to 0}

\item{sig}{Significance parameter, used in dnorm as the final quantile}

\item{kern.trunc}{Truncated value for the kernel}

\item{nw}{A positive double-precision number, the time-bandwidth parameter}
}
\value{
Gaussian matrix scaled for bright
}
\description{
Non bright image scaling with gaussian deconvolution
}
