% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{.process_image}
\alias{.process_image}
\title{Process Image}
\usage{
.process_image(
  imageMatrix,
  beta0 = -2.774327,
  beta1 = 51.91687,
  cutoffProbability = 0.5,
  NADefault = 0
)
}
\arguments{
\item{imageMatrix}{Imported image into matrix form, can use tiff_import()}

\item{beta0}{From logistic regression on what images to be considered bright}

\item{beta1}{From logistic regression on what images to be considered bright}

\item{cutoffProbability}{Passed into bright: The probability cut off for the decision of an imageMatrix being bright}

\item{NADefault}{The defult value set to points of NA found by the system}
}
\value{
The processed image with the gaussian and the non gaussian in an array labeled respectively
}
\description{
Process Image
}
