% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{.trim_Sides}
\alias{.trim_Sides}
\title{Rough Trim for Sides}
\usage{
.trim_Sides(image, trimAmountLeft = 2, trimAmountRight = 2)
}
\arguments{
\item{image}{A matrix}

\item{trimAmountLeft}{in percentage}

\item{trimAmountRight}{in percentage}
}
\value{
the matrix without the specified bounds
}
\description{
trims roughly the right and left sides if the user knows where there will definitely not be any points
}
