% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{.trim_top_bottom}
\alias{.trim_top_bottom}
\title{Rough Trim}
\usage{
.trim_top_bottom(image, trimAmountTop = 100, trimAmountBottom = 50)
}
\arguments{
\item{image}{A matrix}

\item{trimAmountTop}{Number of pixels}

\item{trimAmountBottom}{Number of pixels}
}
\value{
the matrix without the specified bounds
}
\description{
trims roughly the top and bottom if the user knows where there will definitly not be any points
}
