% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFunctions.R
\name{.triple_check}
\alias{.triple_check}
\title{Triple Trace Checking}
\usage{
.triple_check(
  imageMatrix,
  minDistance = 50,
  percentFromEdge = 2,
  threshold = 200
)
}
\arguments{
\item{imageMatrix}{Horizontal image processed}

\item{minDistance}{The min distance aloud between found peaks}

\item{percentFromEdge}{the distance alous from the edge of the image}

\item{threshold}{how close the last three peaks have to be in height +-}
}
\value{
TRUE or FALSE for finding a triple or not
}
\description{
Using find_peaks, checks for three traces at the bottom with the same length
}
