% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userFunctions.R
\name{find_envelopes}
\alias{find_envelopes}
\title{Finding Envelopes for Two Traces}
\usage{
find_envelopes(
  rolledImage,
  imageMatrix,
  bottomcut,
  returnType,
  sepDist = 10,
  max_roc = 25,
  maxNoise = 100
)
}
\arguments{
\item{rolledImage}{product of mean_roll_image()}

\item{imageMatrix}{ImageMatrix from import_process_image()}

\item{bottomcut}{from .Image_cut()}

\item{returnType}{Either "MatrixScaled" used if you are manipulating with imageMatrix,
"plottingScaled" used if you are plotting the imageMatrix with overlay of these lines
"rolledImageScaled" used if you are overlaying the rolled image with these lines}

\item{sepDist}{how far you want the envelope to be below the line you are tracing}

\item{max_roc}{maximum rate of change allowed between two pixels on the line before deemed as noise}

\item{maxNoise}{the length of creating points alowed before considered to be off of the trace}
}
\value{
list of all four envelopes
}
\description{
Finding Envelopes for Two Traces
}
