% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.matrix.R
\name{newPixmapRGB}
\alias{newPixmapRGB}
\title{newPixmapRGB}
\usage{
newPixmapRGB(red = NULL, green = NULL, blue = NULL)
}
\arguments{
\item{red}{colour from newPixmapRGB}

\item{green}{colour from newPixmapRGB}

\item{blue}{colour from newPixmapRGB}
}
\description{
Used inside plot.matrix
}
