% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userFunctions.R
\name{TIS}
\alias{TIS}
\title{TIS: Trace Identification Through Separation}
\usage{
TIS(
  imageName,
  fileLoc,
  pathToWorkingDir = "~/",
  HDVcheck = FALSE,
  plotPNG = TRUE,
  trimAmountTop = 100,
  trimAmountBottom = 50,
  beta0 = -2.774327,
  beta1 = 51.91687,
  cutoffProbability = 0.5,
  NADefault = 0,
  FilterBright = c(13, 13),
  FilterNonBright = c(8, 8),
  methodBright = "delation",
  methodNonBright = "delation",
  thresholdBright = 0.8,
  thresholdNonBright = 0.5,
  trimAmountLeft = 2,
  trimAmountRight = 2,
  peakPercFromEdge = 2,
  peakPercentFromEdgeLeftSide = 25,
  cutPercentage = 2,
  shortestAllowedSeqOfZeros = 25,
  TrippleCheckMinDistance = 50,
  tripleThresholdHeight = 200,
  tripleThresholdDistance = 250,
  threshCutImage = 500,
  OffsetDistanceForEnvelopes = 10,
  maxEnvelopeROC = 35,
  maxNoise = 250,
  envelopeStartEndThreshold = 300,
  intersetionRemoveAmount = 1000,
  CreateTraceThreshold = 5,
  MARange = 6,
  region = 2,
  loopNumber = 4,
  spikeThreshold = 50,
  k = 40
)
}
\arguments{
\item{imageName}{The name of the file}

\item{fileLoc}{Path from ~/ to the dir where the file is}

\item{pathToWorkingDir}{Path of where the plots should be saved to}

\item{HDVcheck}{For Magneto project only, use FALSE otherwise}

\item{plotPNG}{TRUE or FALSE if you want plots saved or just data returned}

\item{trimAmountTop}{Number of pixels off of top of image (usually for common flares)}

\item{trimAmountBottom}{Number of pixels off of bottom of image}

\item{beta0}{The intercept of a logistic regression default is for magnetograms
Logistic regression on the decision if an image can be considered to be bright
based off of a user deciding bright or non bright on a large set of images}

\item{beta1}{The slope of the logistic regression default is for magnetograms}

\item{cutoffProbability}{The probability cut off for the decision of an imageMatrix being bright}

\item{NADefault}{The default value set to NA's found in the matrix}

\item{FilterBright}{Vector specifying the dimensions of the kernel,
which will be used to perform either delation or erosion, such as c(13,13)}

\item{FilterNonBright}{Vector specifying the dimensions of the kernel,
which will be used to perform either delation or erosion, such as c(8,8)}

\item{methodBright}{one of 'delation'(adds to image, making brights brighter),
'erosion' (subtracts from image brights darker)}

\item{methodNonBright}{one of 'delation'(adds to image, making brights brighter),
'erosion' (subtracts from image brights darker)}

\item{thresholdBright}{should be between 0 and 1 for normalized images Default = 0.8
threshold for what can be considered to be 1 and what should be considered to be 0(noise)}

\item{thresholdNonBright}{should be between 0 and 1 for normalized images Default = 0.5
#' threshold for what can be considered to be 1(the trace/ white pixel) and
what should be considered to be 0(noise black pixel)}

\item{trimAmountLeft}{Percentage of the image that is guaranteed to not have
any trace in it(mostly used to remove flares from the scanning process)}

\item{trimAmountRight}{Percentage of the image that is guaranteed to not have
any trace in it(mostly used to remove flares from the scanning process)

For the cut's (removing the timing traces and the writing at the top of the image)
look at rowSums of the matrix to see the peaks for these two parameters}

\item{peakPercFromEdge}{Used in find_peaks if you know there wont be a peak
in a region}

\item{peakPercentFromEdgeLeftSide}{Passed into find peaks, if not specified (NULL), uses
peakPercFromEdge for both left and right sides, if specified, peakPercFromEdge
is defaulted to just the right side of the plot}

\item{cutPercentage}{Usually the min of trimAmountLeft and trimAmountRight}

\item{shortestAllowedSeqOfZeros}{Smallest gap allowed between the timing marks
and the timing marks, before the trace is considered to intersect the timing marks}

\item{TrippleCheckMinDistance}{Minimum distance allowed between found peaks.  Ensures that the user
doesn't have two peaks found for where there is only one peak with two points
at the top (only used for magneto, make large if not needed)}

\item{tripleThresholdHeight}{How different the heights can be for three peaks
to be considered to be a triple set of peaks (only used for magneto, make 0 if not needed)}

\item{tripleThresholdDistance}{How far apart the peaks can be for the them to be
considered to be a set of triples (only used for magneto, make large if not needed)}

\item{threshCutImage}{Used in Triple check as a second way to check for triples,
takes the cut image(not timing or writing and does the find peaks on it, checks for 3 peaks)
Same definition as the tripleThresholdDistance but different value needed because different scale}

\item{OffsetDistanceForEnvelopes}{How far off of the trace you would like the envelope to be
used as a safety net to catch any parts of the trace that might not be detected}

\item{maxEnvelopeROC}{The max difference (rate of change) between any two points before considered to have started
to trace noise (little points or the words) and are not running along the actual trace anymore.}

\item{maxNoise}{The max amount of points created off of the trace before the
envelope can be considered to be not filling in a gap but being off of the
trace entirely, will correct after this number}

\item{envelopeStartEndThreshold}{How far the start and end lines will go
measuring in pixels before guaranteed to be past any noise or writing before the trace
(will only use if no other start points are found)}

\item{intersetionRemoveAmount}{The amount removed of the image (from the right and the left)
before looking for an intersection between the two traces}

\item{CreateTraceThreshold}{The cutoff for the ROC between two points, any
points found during the final line tracing higher then the ROC
will be smoothed with a MA}

\item{MARange}{The amount of points in each direction that the moving average
will look at to calculate the MA on (added to teh region value for the whole MARange)}

\item{region}{The region inside the MA that will be corrected to the MA value}

\item{loopNumber}{The amount of times the MA Smoothing will happen, to guarantee creation of other peaks}

\item{spikeThreshold}{The difference between two pixels considered to look for abnormal spikes in the tracing algorithm,
could be due to a jump from one trace to another (anything above the threshold will be sent to warning)}

\item{k}{See rollMean() in zoo package for details}
}
\value{

}
\description{
Combines all other functions to allow the user to digitize one image.
All parameters are adjustable to allow for attempts to digitize more then one
type of image
}
\details{
Beta 0 and Beta1 are required for the brightness, changes the method for the
image processing, if an image is bright, the parameters change to try and remove
the over exposure so the envelopes can be more accurate around a trace.
}
