% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userFunctions.R
\name{TIS_automation}
\alias{TIS_automation}
\title{TIS_automation}
\usage{
TIS_automation(
  DigitizationTODO,
  pathToDigitizationDir,
  keywordInImageName,
  ignoreErrorMessage = FALSE
)
}
\value{

}
\description{
TIS_automation
}
