% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userFunctions.R
\name{find_envelopes}
\alias{find_envelopes}
\title{Finding Envelopes for Two Traces}
\usage{
find_envelopes(
  imageMatrix,
  rolledImage,
  bottomCut,
  returnType,
  sepDist = 10,
  max_roc = 25,
  maxNoise = 100,
  improveTTopEnvelope = NA,
  improveBTopEnvelope = NA,
  improveTBottomEnvelope = NA,
  improveBBottomEnvelope = NA
)
}
\arguments{
\item{imageMatrix}{ImageMatrix from import_process_image()}

\item{rolledImage}{the image rolled by mean_roll_image()}

\item{bottomCut}{a point that is the height of a horizontal line between the timing marks
and the bottom trace}

\item{returnType}{Either "MatrixScaled" used if you are manipulating with imageMatrix,
"plottingScaled" used if you are plotting the imageMatrix with overlay of these lines
"rolledImageScaled" used if you are overlaying the rolled image with these lines}

\item{sepDist}{how far you want the envelope to be below the line you are tracing}

\item{max_roc}{maximum rate of change allowed between two pixels on the line before deemed as noise}

\item{maxNoise}{the length of creating points allowed before considered to be off of the trace}
}
\value{
list of all four envelopes
}
\description{
Finding Envelopes for Two Traces
}
